+++ .gitignore
@@ -12,7 +12,7 @@ logs/
 
 # Directories
 coverage/
-build/
+/build/
 cjs/
 dist/
 dts/
